refactor: Modularize browser layout into separate components
Extracted BrowserHeader, BrowserHeroSection, and FeaturedWebsites into separate components

Updated IconBrowser usage to align with the new structure.

Refactored SearchOverlay and index.ts to accommodate the new components.
