+++ lib/components/browser/index.ts
@@ -1,4 +1,7 @@
+import BrowserHeader from "./BrowserHeader";
+import BrowserHeroSection from "./BrowserHeroSection";
+import FeaturedWebsites from "./FeaturedWebsites";
 import QuickActions from "./QuickActions";
 import SearchOverlay from "./SearchOverlay";
 
-export { SearchOverlay, QuickActions };
+export { BrowserHeader, BrowserHeroSection, FeaturedWebsites, SearchOverlay, QuickActions };
